More major refactoring.
