+++ src/Utilities.cpp
@@ -1,0 +1,12 @@
+/**
+ * @file Utilities.cpp
+ * @author Richard Buckley <richard.buckley@ieee.org>
+ * @version 1.0
+ * @date 2021-03-06
+ */
+
+#include "Utilities.h"
+
+namespace rose {
+
+}
